requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
sqlalchemy==2.0.35
alembic==1.13.2
pydantic==2.9.2
psutil==6.0.0
scikit-learn==1.5.2
pandas==2.2.3
numpy==2.1.1
python-dotenv==1.0.1
celery==5.4.0
redis==5.2.1
fastapi==0.115.0
uvicorn==0.30.6
sqlalchemy==2.0.35
pydantic==2.9.2
pydantic-settings==2.5.2
psutil==6.0.0
scikit-learn==1.5.2
pandas==2.2.3
numpy==2.1.1
python-dotenv==1.0.1
apscheduler==3.10.4
celery==5.4.0
redis==5.2.1        
